TINY-8312: Remove dupe and use alien files instead
