+++ modules/tinymce/src/core/main/ts/textpatterns/keyboard/KeyHandler.ts
@@ -7,13 +7,13 @@
 
 import { Unicode } from '@ephox/katamari';
 
+import { textBefore } from 'tinymce/core/alien/TextSearch';
 import Editor from 'tinymce/core/api/Editor';
 import VK from 'tinymce/core/api/util/VK';
 
 import * as BlockPattern from '../core/BlockPattern';
 import * as InlinePattern from '../core/InlinePattern';
 import { PatternSet } from '../core/PatternTypes';
-import { textBefore } from '../text/TextSearch';
 import { cleanEmptyNodes } from '../utils/Utils';
 
 const handleEnter = (editor: Editor, patternSet: PatternSet): boolean => {
